CR fixes

+++ contracts/ProAMMPoolTicksState.sol
@@ -26,8 +26,8 @@ contract ProAMMPoolTicksState is IProAMMPoolTicksState {
     // seconds spent on the other side of this tick (relative to current tick)
     // only has relative meaning, not absolute — the value depends on when the tick is initialized
     uint160 secondsPerLiquidityOutside;
-    // true iff the tick is initialized, when liquidityGross != 0
-    // these 8 bits are set to prevent fresh sstores when crossing newly initialized ticks
+    // true if liquidityGross != 0, false otherwise
+    // this prevents fresh sstores when crossing newly initialized ticks
     bool initialized;
   }
 
@@ -51,7 +51,7 @@ contract ProAMMPoolTicksState is IProAMMPoolTicksState {
 
   struct CumulativesData {
     uint256 feeGrowth;
-    uint160 secondsPerLiquidity;
+    uint128 secondsPerLiquidity;
   }
 
   // uint128 public immutable maxLiquidityPerTick;
